# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Smart Recruitment System API
  version: 1.0.0
  description: |
    REST API لمنظومة التوظيف الذكية. المصادقة عبر Bearer JWT.
    الواجهة الحالية تحاكي كل هذه المسارات محليًا (Pinia + localStorage)؛
    التبديل للحقيقي عبر VITE_USE_REAL_API=true دون تغيير واجهات المخازن.
servers:
  - url: http://localhost:8000/api/v1
    description: بيئة التطوير المحلية (Laravel)
security:
  - bearerAuth: []
paths:
...
  /public-profiles/{slug}:
    get:
      tags: [PublicProfile]
      security: []
      summary: الصفحة العامة لمستخدم (للزوار — بلا مصادقة)
      parameters: [{ $ref: '#/components/parameters/slug' }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicProfile' }
        '404': { $ref: '#/components/responses/Error' }
  /public-profiles/me:
    get:
      tags: [PublicProfile]
      summary: صفحتي للتحرير/الإماهة (get-or-create — نفس شكل الحالة المحلية)
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicProfile' }
    patch:
      tags: [PublicProfile]
      summary: تحديث صفحتي (الهوية/المظهر/الأقسام/المحتوى — نفس شكل الحالة المحلية)
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublicProfile' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicProfile' }
  /public-profiles/{slug}/view:
    post:
      tags: [PublicProfile]
      security: []
      summary: تسجيل مشاهدة
      parameters: [{ $ref: '#/components/parameters/slug' }]
      responses:
        '204': { description: سُجّلت }
  /public-profiles/{slug}/follow:
    post:
      tags: [PublicProfile]
      security: []
      summary: متابعة/إلغاء متابعة (تبديل)
      parameters: [{ $ref: '#/components/parameters/slug' }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  following: { type: boolean }
                  followersCount: { type: integer }
  /public-profiles/{slug}/rate:
    post:
      tags: [PublicProfile]
      security: []
      summary: تقييم الصفحة (تعديل التقييم لا يضاعف العداد)
      parameters: [{ $ref: '#/components/parameters/slug' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [stars]
              properties:
                stars: { type: integer, minimum: 1, maximum: 5 }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  avgRating: { type: number }
                  ratingCount: { type: integer }
  /public-profiles/{slug}/comments:
    post:
      tags: [PublicProfile]
      security: []
      summary: تعليق زائر (يخضع لإشراف صاحب الصفحة)
      parameters: [{ $ref: '#/components/parameters/slug' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [author, text]
              properties:
                author: { type: string }
                text: { type: string }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageComment' }
  /public-profiles/{slug}/contact:
    post:
      tags: [PublicProfile]
      security: []
      summary: «تواصل معي» — تصل رسائل صاحب الصفحة داخل المنصة
      parameters: [{ $ref: '#/components/parameters/slug' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [visitorName, text]
              properties:
                visitorName: { type: string }
                text: { type: string }
      responses:
        '204': { description: وصلت }
  /public-profiles/{slug}/schedule:
    post:
      tags: [PublicProfile]
      security: []
      summary: اقتراح موعد مقابلة من زائر
      parameters: [{ $ref: '#/components/parameters/slug' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [visitorName, day, slot]
              properties:
                visitorName: { type: string }
                day: { type: string }
                slot: { type: string }
                topic: { type: string }
      responses:
        '204': { description: وصل الاقتراح }
  /public-profiles/{slug}/testimonials:
    post:
      tags: [PublicProfile]
      security: []
      summary: توصية من زائر (تدخل مخفية حتى موافقة صاحب الصفحة)
      parameters: [{ $ref: '#/components/parameters/slug' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [author, excerpt]
              properties:
                author: { type: string }
                authorRole: { type: string }
                excerpt: { type: string }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Testimonial' }
  /public-profiles/{slug}/proof-requests:
    post:
      tags: [PublicProfile]
      security: []
      summary: زائر يطلب إثبات مهارة من صاحب الصفحة
      parameters: [{ $ref: '#/components/parameters/slug' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [skill, from]
              properties:
                skill: { type: string }
                from: { type: string }
                relation: { type: string }
      responses:
        '204': { description: وصل الطلب }

  # ===== الفرص وسوق الطلبات =====
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    slug:
      name: slug
      in: path
      required: true
      schema: { type: string }
  responses:
...
    Error:
      description: خطأ
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
  schemas:
    ApiError:
      type: object
      properties:
        message: { type: string }
        code: { type: string }
...
    PublicProfile:
      type: object
      description: نفس شكل PublicProfileState في PublicProfileStore.ts (المصدر المرجعي)
      properties:
        slug: { type: string }
        displayName: { type: string }
        publicHeadline: { type: string }
        tagline: { type: string }
        bioShort: { type: string }
        story: { type: string }
        keywords:
          type: array
          items: { type: string }
        location: { type: string }
        timezone: { type: string }
        avatarImage: { type: [string, 'null'], description: رابط صورة (الخادم يستضيفها بدل dataURL) }
        availability:
          type: object
          properties:
            status: { type: string, enum: [available, busy, unavailable] }
            message: { type: string }
        appearance:
          type: object
          description: الثيم والخط وشكل الصورة والحركة — كما في ProfileAppearance
          additionalProperties: true
        sectionOrder:
          type: array
          items: { type: string }
        sections:
          type: object
          additionalProperties: { type: boolean }
        links:
          type: object
          additionalProperties: { type: string }
        customLinks:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              label: { type: string }
              url: { type: string }
        achievements:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              text: { type: string }
              kind: { type: string, enum: [self, verified] }
        portfolio:
          type: array
          items: { $ref: '#/components/schemas/PortfolioItem' }
        testimonials:
          type: array
          items: { $ref: '#/components/schemas/Testimonial' }
        stats:
          type: object
          description: عدادات الجذب (قراءة فقط للمالك)
          properties:
            views: { type: integer }
            shares: { type: integer }
            contacts: { type: integer }
            meetings: { type: integer }
            followersCount: { type: integer }
            avgRating: { type: number }
            ratingCount: { type: integer }
    PortfolioItem:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        desc: { type: string }
        tag: { type: string }
        link: { type: string }
        image: { type: string, description: رابط صورة مستضافة }
    Testimonial:
      type: object
      properties:
        id: { type: integer }
        author: { type: string }
        authorRole: { type: string }
        excerpt: { type: string }
        visible: { type: boolean }
        likes: { type: integer }
    PageComment:
      type: object
      properties:
        id: { type: integer }
        author: { type: string }
        text: { type: string }
        date: { type: string, format: date }
        hidden: { type: boolean }
...